AuthGuard2/requirements.txt
Flask>=2.0
flask-cors>=3.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0